+++ diary-app-frontend/src/components/DiaryList.js
@@ -6,8 +6,10 @@ import diaryService from "../services/diaryService";
 // 日記を表示するコンポーネント
 const DiaryList = ({ diaries }) => {
   
-  //削除ボタンを押下したときの処理
+    //削除ボタンを押下したときの処理
     const deleteDiary = (id) => {
+      const result = window.confirm("削除します。本当に宜しいですか？");
+      if (result) {
         diaryService.deleteDiary(id).then(() => {
           alert("指定された日記を削除しました");
           //TODO:一覧画面に遷移する処理をここに
@@ -16,6 +18,7 @@ const DiaryList = ({ diaries }) => {
             console.error("Error delete diary: ", err);
             alert("指定された日記の削除に失敗しました");
         })
+      }
     }
 
   const navigate = useNavigate();
